// src/dashboard/types/enums.ts
// Enum for User Gender
export enum UserGender {
  Male = "male",
  Female = "female",
}

// Enum for Target Group
export enum TargetGroup {
  MaleOnly = "male_only",
  FemaleOnly = "female_only",
  All = "all",
}

export enum VerificationStatus {
  Verified = "verified",
  Unverified = "unverified",
  Pending = "pending",
}

export enum AccountStatus {
  Active = "active",
  Inactive = "inactive",
  Suspended = "suspended",
}

export enum AccountType {
  Business = "business",
  User = "user",
}

export enum Gender {
  Male = "male",
  Female = "female",
}

export enum BasicAllScaapeFilter {
  ALL = "all",
  NEAR_ME = "near_me",
  RECOMMENDED = "recommended",
  POPULAR = "popular",
  PREVIOUSLY_ATTENDED = "previously_attended",
}

export enum ScaapeParticipantsStatus {
  Pending = "pending",
  Accepted = "accepted",
  Rejected = "rejected",
}